LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/static-router-multiple-router-registrations.https-expected.txt


FAIL Main reosurce load matched with the service worker having multiple rules assert_equals: expected "Network\n" but got ""
FAIL Resource load matched with the rule registered in the imported service worker assert_equals: expected "Network\n" but got ""
